feat: Implement full SIMD acceleration for LZ4 and Zstd
## Changes

### SIMD Implementations
- LZ4 AVX-512: Full 512-bit decompression with 64-byte copies
- LZ4 NEON: Full 128-bit decompression for ARM/Apple Silicon
- Zstd AVX2: Huffman/FSE decoding with VPSHUFB table lookups
- Runtime SIMD dispatch: decompress_simd() auto-selects best backend

### Dependency Updates
- thiserror: 1.0 → 2.0.17
- clap: 4.4 → 4.5.54
- tokio: 1.35 → 1.49.0
- proptest: 1.4 → 1.6

### Quality Improvements
- SATD violations: 6 → 0 (all TODOs implemented)
- Dead code: 1.46% → 0.09%
- Tests: 73 → 95 passing
- Coverage: 71.11% (limited by SIMD feature availability)

## Performance Targets
- LZ4 AVX-512: ≥5 GB/s decompression
- LZ4 AVX2/NEON: ≥4 GB/s decompression
- LZ4 compression: ≥3 GB/s (hash-table bound)

+++ crates/trueno-zram-core/src/zstd/decompress.rs
@@ -137,8 +137,8 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
         }
         let wd = input[*pos];
         *pos += 1;
-        let exponent = (wd >> 3) as u64;
-        let mantissa = (wd & 0x07) as u64;
+        let exponent = u64::from(wd >> 3);
+        let mantissa = u64::from(wd & 0x07);
         (1u64 << (10 + exponent)) + (mantissa << (7 + exponent))
     };
 
@@ -149,7 +149,7 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
             if *pos >= input.len() {
                 return Err(Error::CorruptedData("missing dict id".to_string()));
             }
-            let id = input[*pos] as u32;
+            let id = u32::from(input[*pos]);
             *pos += 1;
             Some(id)
         }
@@ -157,7 +157,7 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
             if *pos + 2 > input.len() {
                 return Err(Error::CorruptedData("missing dict id".to_string()));
             }
-            let id = u16::from_le_bytes([input[*pos], input[*pos + 1]]) as u32;
+            let id = u32::from(u16::from_le_bytes([input[*pos], input[*pos + 1]]));
             *pos += 2;
             Some(id)
         }
@@ -185,7 +185,7 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
                     "missing frame content size".to_string(),
                 ));
             }
-            let size = input[*pos] as u64;
+            let size = u64::from(input[*pos]);
             *pos += 1;
             Some(size)
         }
@@ -196,7 +196,7 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
                     "missing frame content size".to_string(),
                 ));
             }
-            let size = u16::from_le_bytes([input[*pos], input[*pos + 1]]) as u64 + 256;
+            let size = u64::from(u16::from_le_bytes([input[*pos], input[*pos + 1]])) + 256;
             *pos += 2;
             Some(size)
         }
@@ -206,12 +206,12 @@ fn read_frame_header(input: &[u8], pos: &mut usize) -> Result<FrameHeader> {
                     "missing frame content size".to_string(),
                 ));
             }
-            let size = u32::from_le_bytes([
+            let size = u64::from(u32::from_le_bytes([
                 input[*pos],
                 input[*pos + 1],
                 input[*pos + 2],
                 input[*pos + 3],
-            ]) as u64;
+            ]));
             *pos += 4;
             Some(size)
         }
@@ -252,7 +252,7 @@ fn read_block_header(input: &[u8], pos: &mut usize) -> Result<BlockHeader> {
     }
 
     let header =
-        (input[*pos] as u32) | ((input[*pos + 1] as u32) << 8) | ((input[*pos + 2] as u32) << 16);
+        u32::from(input[*pos]) | (u32::from(input[*pos + 1]) << 8) | (u32::from(input[*pos + 2]) << 16);
     *pos += 3;
 
     let last_block = (header & 0x01) != 0;
